Fix int overflow in feature extraction (ref #237).
Former-commit-id: fdd2dd744449d226b213cd48691fc4edbee4c235

+++ palladian-core/src/main/java/ws/palladian/extraction/location/disambiguation/LocationFeatureExtractor.java
@@ -404,12 +404,12 @@ class LocationFeatureExtractor {
         return LocationExtractorUtils.largestDistanceBelow(50, coordinates);
     }
 
-    private static int getPopulationInRadius(Location location, Collection<Location> others, double distance) {
+    private static long getPopulationInRadius(Location location, Collection<Location> others, double distance) {
         GeoCoordinate locationCoordinate = location.getCoordinate();
         if (locationCoordinate == null) {
             return 0;
         }
-        int population = 0;
+        long population = 0;
 //        if (locationCoordinate != null) {
 //            for (Location other : others) {
 //                GeoCoordinate otherCoordinate = other.getCoordinate();
